Add files via upload
The demographic analysis revealed that the majority of respondents (80%) were between the ages of 20–24, which aligns with the typical undergraduate age range in Nigeria. Over 90% of participants were single, which is expected given the student population. In terms of religious affiliation, Christianity emerged as the dominant religion among the students, followed by Islam. When asked about the age of sexual debut, most respondents (46%) reported initiating sexual activity between the ages of 16–20, indicating relatively early exposure to sexual behavior.

Several key factors were identified as influencing contraceptive use among the participants. Fear of unwanted pregnancy stood out as a significant motivator, with 93% of respondents indicating it as a reason for using contraceptives. Similarly, 92% cited the fear of contracting sexually transmitted infections (STIs) as another major influencing factor. Partner dynamics also played a crucial role, as 73% of respondents mentioned that their partner’s insistence encouraged the use of contraceptives. Additionally, religious beliefs were acknowledged by around 52.5% of the respondents as having an impact on their decision to use, or not use, contraceptive methods.